+++ src/variational_parameters.cpp
@@ -228,7 +228,7 @@ VariationalParametersLite VariationalParameters::convert_to_lite() const {
 	return vplite;
 }
 
-void VariationalParameters::calcEdZtZ(const Eigen::Ref<const Eigen::ArrayXXd> &dXtEEX_lowertri, const int &n_env) {
+void VariationalParameters::calcEdZtZ(const Eigen::Ref<const Eigen::ArrayXXd> &dXtEEX_lowertri, const long &n_env) {
 	Eigen::ArrayXd muw_sq_combos(n_env * (n_env + 1) / 2);
 	for (int ll = 0; ll < n_env; ll++) {
 		for (int mm = 0; mm < n_env; mm++) {
